Fix double rounding errors for disk usage

+++ server/src/main/java/org/elasticsearch/cluster/routing/allocation/decider/DiskThresholdDecider.java
@@ -595,7 +595,7 @@ public class DiskThresholdDecider extends AllocationDecider {
             if (getTotalBytes() == 0L) {
                 return 100.0;
             }
-            return 100.0 * ((double) getFreeBytes() / getTotalBytes());
+            return 100.0 * getFreeBytes() / getTotalBytes();
         }
 
         double getUsedDiskAsPercentage() {
